+++ electron/main.ts
@@ -431,6 +431,7 @@ async function initServer() {
         createErrorLog: s.createErrorLog,
         savePlaylistAsCompilation: s.savePlaylistAsCompilation,
         createPlaylistFile: s.createPlaylistFile,
+        createAlbumPlaylistFile: s.createAlbumPlaylistFile,
         overwriteFiles: s.overwriteFiles,
         skipDuplicateTracks: s.skipDuplicateTracks,
         maxConcurrentDownloads: s.maxConcurrentDownloads,
@@ -441,6 +442,7 @@ async function initServer() {
           createCDFolder: s.createCDFolder,
           createPlaylistStructure: s.createPlaylistStructure,
           createSinglesStructure: s.createSinglesStructure,
+          createShortReleaseFolder: s.createShortReleaseFolder,
           playlistFolderTemplate: s.playlistFolderTemplate,
           albumFolderTemplate: s.albumFolderTemplate,
           artistFolderTemplate: s.artistFolderTemplate
@@ -520,6 +522,7 @@ async function initServer() {
           createCDFolder: s.createCDFolder,
           createPlaylistStructure: s.createPlaylistStructure,
           createSinglesStructure: s.createSinglesStructure,
+          createShortReleaseFolder: s.createShortReleaseFolder,
           playlistFolderTemplate: s.playlistFolderTemplate,
           albumFolderTemplate: s.albumFolderTemplate,
           artistFolderTemplate: s.artistFolderTemplate